chore: clean up code and rename file
